% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cescalate-package.R
\docType{package}
\name{cescalate-package}
\alias{cescalate}
\alias{cescalate-package}
\title{cescalate: Exploring toxicity/efficacy dose-escalation designs with continuous efficacy}
\description{
This package performs model fitting, simulation, and visualization for various efficacy and toxicity interval design for phase I/II dose finding trials such as checkerboard by Yin and Yuan (10.1080/10543406.2020.1815033)
}
\author{
\strong{Maintainer}: Quang Nguyen \email{quangpmnguyen@gmail.com}

}
